refactor: rename plugins and flatten dirs
