+++ src/build.ts
@@ -1,9 +1,7 @@
 import * as esbuild from "https://deno.land/x/esbuild@v0.17.16/mod.js";
-import { island_wrapper } from "./esbuild_plugins/islands.ts";
-import {
-	resolve_svelte_internal,
-} from "./esbuild_plugins/resolve_svelte_internal.ts";
-import { build_routes } from "./esbuild_plugins/build_routes/index.ts";
+import { svelte_components } from "./esbuild_plugins/svelte_components.ts";
+import { svelte_internal } from "./esbuild_plugins/svelte_internal.ts";
+import { build_routes } from "./esbuild_plugins/routes.ts";
 import { ensureDir } from "https://deno.land/std@0.177.0/fs/ensure_dir.ts";
 import { parse } from "https://deno.land/std@0.177.0/flags/mod.ts";
 import { serve } from "https://deno.land/std@0.177.0/http/server.ts";
@@ -63,8 +61,8 @@ const routesESBuildConfig: esbuild.BuildOptions = {
 	entryPoints: await get_svelte_files({ dir: "routes/" }),
 	write: false,
 	plugins: [
-		island_wrapper(site_dir),
-		resolve_svelte_internal,
+		svelte_components(site_dir),
+		svelte_internal,
 		build_routes({ base_path }),
 	],
 	outdir: build_dir,
@@ -75,8 +73,8 @@ const islandsESBuildConfig: esbuild.BuildOptions = {
 	entryPoints: await get_svelte_files({ dir: "components/" }),
 	write: true,
 	plugins: [
-		island_wrapper(site_dir),
-		resolve_svelte_internal,
+		svelte_components(site_dir),
+		svelte_internal,
 	],
 	outdir: build_dir + "components/",
 	splitting: true,
