+++ src/esbuild_plugins/svelte_internal.ts
@@ -1,11 +1,10 @@
 import type { Plugin } from "https://deno.land/x/esbuild@v0.17.16/mod.js";
 
-const svelte_internal = await fetch(
+const contents = await fetch(
 	"https://cdn.jsdelivr.net/npm/svelte@3.58.0/internal/index.mjs",
-);
-const svelte_internal_src = await svelte_internal.text();
+).then((r) => r.text());
 
-export const resolve_svelte_internal: Plugin = {
+export const svelte_internal: Plugin = {
 	name: "svelte/internal",
 	setup(build) {
 		build.onResolve({ filter: /^svelte(\/internal)?$/ }, () => {
@@ -18,7 +17,7 @@ export const resolve_svelte_internal: Plugin = {
 
 		build.onLoad({ filter: /.*/, namespace: "svelte" }, () => {
 			return {
-				contents: svelte_internal_src,
+				contents,
 			};
 		});
 	},
